# kgs-platform/api/graph/v1/graph_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v4.25.1
// source: graph/v1/graph.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationGraphCreateEdge = "/api.graph.v1.Graph/CreateEdge"
const OperationGraphCreateNode = "/api.graph.v1.Graph/CreateNode"
const OperationGraphDeleteNode = "/api.graph.v1.Graph/DeleteNode"
const OperationGraphGetContext = "/api.graph.v1.Graph/GetContext"
const OperationGraphGetCoverage = "/api.graph.v1.Graph/GetCoverage"
const OperationGraphGetImpact = "/api.graph.v1.Graph/GetImpact"
const OperationGraphGetNode = "/api.graph.v1.Graph/GetNode"
const OperationGraphGetSubgraph = "/api.graph.v1.Graph/GetSubgraph"
const OperationGraphUpdateNode = "/api.graph.v1.Graph/UpdateNode"

type GraphHTTPServer interface {
	// CreateEdge Create an edge between two nodes
	CreateEdge(context.Context, *CreateEdgeRequest) (*CreateEdgeReply, error)
	// CreateNode Create a new node in the Knowledge Graph
	CreateNode(context.Context, *CreateNodeRequest) (*CreateNodeReply, error)
	// DeleteNode Delete a node
	DeleteNode(context.Context, *DeleteNodeRequest) (*DeleteNodeReply, error)
	// GetContext Get contextual neighborhood around a node
	GetContext(context.Context, *GetContextRequest) (*GraphReply, error)
	// GetCoverage Get upstream coverage of a node
	GetCoverage(context.Context, *GetCoverageRequest) (*GraphReply, error)
	// GetImpact Get downstream impact of a node
	GetImpact(context.Context, *GetImpactRequest) (*GraphReply, error)
	// GetNode Get a node by ID
	GetNode(context.Context, *GetNodeRequest) (*GetNodeReply, error)
	// GetSubgraph Get subgraph for a list of node IDs
	GetSubgraph(context.Context, *GetSubgraphRequest) (*GraphReply, error)
	// UpdateNode Update a node's properties
	UpdateNode(context.Context, *UpdateNodeRequest) (*UpdateNodeReply, error)
}

func RegisterGraphHTTPServer(s *http.Server, srv GraphHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/graph/nodes", _Graph_CreateNode0_HTTP_Handler(srv))
	r.GET("/v1/graph/nodes/{node_id}", _Graph_GetNode0_HTTP_Handler(srv))
	r.POST("/v1/graph/edges", _Graph_CreateEdge0_HTTP_Handler(srv))
	r.PUT("/v1/graph/nodes/{node_id}", _Graph_UpdateNode0_HTTP_Handler(srv))
	r.DELETE("/v1/graph/nodes/{node_id}", _Graph_DeleteNode0_HTTP_Handler(srv))
	r.GET("/v1/graph/nodes/{node_id}/context", _Graph_GetContext0_HTTP_Handler(srv))
	r.GET("/v1/graph/nodes/{node_id}/impact", _Graph_GetImpact0_HTTP_Handler(srv))
	r.GET("/v1/graph/nodes/{node_id}/coverage", _Graph_GetCoverage0_HTTP_Handler(srv))
	r.POST("/v1/graph/subgraph", _Graph_GetSubgraph0_HTTP_Handler(srv))
}

func _Graph_CreateNode0_HTTP_Handler(srv GraphHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateNodeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGraphCreateNode)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateNode(ctx, req.(*CreateNodeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateNodeReply)
		return ctx.Result(200, reply)
	}
}

func _Graph_GetNode0_HTTP_Handler(srv GraphHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetNodeRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGraphGetNode)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetNode(ctx, req.(*GetNodeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetNodeReply)
		return ctx.Result(200, reply)
	}
}

func _Graph_CreateEdge0_HTTP_Handler(srv GraphHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateEdgeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGraphCreateEdge)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateEdge(ctx, req.(*CreateEdgeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateEdgeReply)
		return ctx.Result(200, reply)
	}
}

func _Graph_UpdateNode0_HTTP_Handler(srv GraphHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateNodeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGraphUpdateNode)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateNode(ctx, req.(*UpdateNodeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateNodeReply)
		return ctx.Result(200, reply)
	}
}

func _Graph_DeleteNode0_HTTP_Handler(srv GraphHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteNodeRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGraphDeleteNode)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteNode(ctx, req.(*DeleteNodeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteNodeReply)
		return ctx.Result(200, reply)
	}
}

func _Graph_GetContext0_HTTP_Handler(srv GraphHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetContextRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGraphGetContext)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetContext(ctx, req.(*GetContextRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GraphReply)
		return ctx.Result(200, reply)
	}
}

func _Graph_GetImpact0_HTTP_Handler(srv GraphHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetImpactRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGraphGetImpact)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetImpact(ctx, req.(*GetImpactRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GraphReply)
		return ctx.Result(200, reply)
	}
}

func _Graph_GetCoverage0_HTTP_Handler(srv GraphHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetCoverageRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGraphGetCoverage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetCoverage(ctx, req.(*GetCoverageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GraphReply)
		return ctx.Result(200, reply)
	}
}

func _Graph_GetSubgraph0_HTTP_Handler(srv GraphHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetSubgraphRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationGraphGetSubgraph)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetSubgraph(ctx, req.(*GetSubgraphRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GraphReply)
		return ctx.Result(200, reply)
	}
}

type GraphHTTPClient interface {
	// CreateEdge Create an edge between two nodes
	CreateEdge(ctx context.Context, req *CreateEdgeRequest, opts ...http.CallOption) (rsp *CreateEdgeReply, err error)
	// CreateNode Create a new node in the Knowledge Graph
	CreateNode(ctx context.Context, req *CreateNodeRequest, opts ...http.CallOption) (rsp *CreateNodeReply, err error)
	// DeleteNode Delete a node
	DeleteNode(ctx context.Context, req *DeleteNodeRequest, opts ...http.CallOption) (rsp *DeleteNodeReply, err error)
	// GetContext Get contextual neighborhood around a node
	GetContext(ctx context.Context, req *GetContextRequest, opts ...http.CallOption) (rsp *GraphReply, err error)
	// GetCoverage Get upstream coverage of a node
	GetCoverage(ctx context.Context, req *GetCoverageRequest, opts ...http.CallOption) (rsp *GraphReply, err error)
	// GetImpact Get downstream impact of a node
	GetImpact(ctx context.Context, req *GetImpactRequest, opts ...http.CallOption) (rsp *GraphReply, err error)
	// GetNode Get a node by ID
	GetNode(ctx context.Context, req *GetNodeRequest, opts ...http.CallOption) (rsp *GetNodeReply, err error)
	// GetSubgraph Get subgraph for a list of node IDs
	GetSubgraph(ctx context.Context, req *GetSubgraphRequest, opts ...http.CallOption) (rsp *GraphReply, err error)
	// UpdateNode Update a node's properties
	UpdateNode(ctx context.Context, req *UpdateNodeRequest, opts ...http.CallOption) (rsp *UpdateNodeReply, err error)
}

type GraphHTTPClientImpl struct {
	cc *http.Client
}

func NewGraphHTTPClient(client *http.Client) GraphHTTPClient {
	return &GraphHTTPClientImpl{client}
}

// CreateEdge Create an edge between two nodes
func (c *GraphHTTPClientImpl) CreateEdge(ctx context.Context, in *CreateEdgeRequest, opts ...http.CallOption) (*CreateEdgeReply, error) {
	var out CreateEdgeReply
	pattern := "/v1/graph/edges"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationGraphCreateEdge))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// CreateNode Create a new node in the Knowledge Graph
func (c *GraphHTTPClientImpl) CreateNode(ctx context.Context, in *CreateNodeRequest, opts ...http.CallOption) (*CreateNodeReply, error) {
	var out CreateNodeReply
	pattern := "/v1/graph/nodes"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationGraphCreateNode))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteNode Delete a node
func (c *GraphHTTPClientImpl) DeleteNode(ctx context.Context, in *DeleteNodeRequest, opts ...http.CallOption) (*DeleteNodeReply, error) {
	var out DeleteNodeReply
	pattern := "/v1/graph/nodes/{node_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGraphDeleteNode))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetContext Get contextual neighborhood around a node
func (c *GraphHTTPClientImpl) GetContext(ctx context.Context, in *GetContextRequest, opts ...http.CallOption) (*GraphReply, error) {
	var out GraphReply
	pattern := "/v1/graph/nodes/{node_id}/context"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGraphGetContext))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetCoverage Get upstream coverage of a node
func (c *GraphHTTPClientImpl) GetCoverage(ctx context.Context, in *GetCoverageRequest, opts ...http.CallOption) (*GraphReply, error) {
	var out GraphReply
	pattern := "/v1/graph/nodes/{node_id}/coverage"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGraphGetCoverage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetImpact Get downstream impact of a node
func (c *GraphHTTPClientImpl) GetImpact(ctx context.Context, in *GetImpactRequest, opts ...http.CallOption) (*GraphReply, error) {
	var out GraphReply
	pattern := "/v1/graph/nodes/{node_id}/impact"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGraphGetImpact))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetNode Get a node by ID
func (c *GraphHTTPClientImpl) GetNode(ctx context.Context, in *GetNodeRequest, opts ...http.CallOption) (*GetNodeReply, error) {
	var out GetNodeReply
	pattern := "/v1/graph/nodes/{node_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationGraphGetNode))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetSubgraph Get subgraph for a list of node IDs
func (c *GraphHTTPClientImpl) GetSubgraph(ctx context.Context, in *GetSubgraphRequest, opts ...http.CallOption) (*GraphReply, error) {
	var out GraphReply
	pattern := "/v1/graph/subgraph"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationGraphGetSubgraph))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateNode Update a node's properties
func (c *GraphHTTPClientImpl) UpdateNode(ctx context.Context, in *UpdateNodeRequest, opts ...http.CallOption) (*UpdateNodeReply, error) {
	var out UpdateNodeReply
	pattern := "/v1/graph/nodes/{node_id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationGraphUpdateNode))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}