# kgs-platform/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/graph/edges:
        post:
            tags:
                - Graph
            description: Create an edge between two nodes
            operationId: Graph_CreateEdge
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.graph.v1.CreateEdgeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.graph.v1.CreateEdgeReply'
    /v1/graph/nodes:
        post:
            tags:
                - Graph
            description: Create a new node in the Knowledge Graph
            operationId: Graph_CreateNode
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.graph.v1.CreateNodeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.graph.v1.CreateNodeReply'
    /v1/graph/nodes/{nodeId}:
        get:
            tags:
                - Graph
            description: Get a node by ID
            operationId: Graph_GetNode
            parameters:
                - name: nodeId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.graph.v1.GetNodeReply'
        put:
            tags:
                - Graph
            description: Update a node's properties
            operationId: Graph_UpdateNode
            parameters:
                - name: nodeId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.graph.v1.UpdateNodeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.graph.v1.UpdateNodeReply'
        delete:
            tags:
                - Graph
            description: Delete a node
            operationId: Graph_DeleteNode
            parameters:
                - name: nodeId
                  in: path
                  required: true
                  schema:
                    type: string
                - name: appId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.graph.v1.DeleteNodeReply'
    /v1/graph/nodes/{nodeId}/context:
        get:
            tags:
                - Graph
            description: Get contextual neighborhood around a node
            operationId: Graph_GetContext
            parameters:
                - name: nodeId
                  in: path
                  required: true
                  schema:
                    type: string
                - name: depth
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: direction
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.graph.v1.GraphReply'
    /v1/graph/nodes/{nodeId}/coverage:
        get:
            tags:
                - Graph
            description: Get upstream coverage of a node
            operationId: Graph_GetCoverage
            parameters:
                - name: nodeId
                  in: path
                  required: true
                  schema:
                    type: string
                - name: maxDepth
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.graph.v1.GraphReply'
    /v1/graph/nodes/{nodeId}/impact:
        get:
            tags:
                - Graph
            description: Get downstream impact of a node
            operationId: Graph_GetImpact
            parameters:
                - name: nodeId
                  in: path
                  required: true
                  schema:
                    type: string
                - name: maxDepth
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.graph.v1.GraphReply'
    /v1/graph/subgraph:
        post:
            tags:
                - Graph
            description: Get subgraph for a list of node IDs
            operationId: Graph_GetSubgraph
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.graph.v1.GetSubgraphRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.graph.v1.GraphReply'
...
components:
    schemas:
...
        api.graph.v1.CreateEdgeReply:
            type: object
            properties:
                edgeId:
                    type: string
                sourceNodeId:
                    type: string
                targetNodeId:
                    type: string
                relationType:
                    type: string
                propertiesJson:
                    type: string
        api.graph.v1.CreateEdgeRequest:
            type: object
            properties:
                sourceNodeId:
                    type: string
                targetNodeId:
                    type: string
                relationType:
                    type: string
                propertiesJson:
                    type: string
        api.graph.v1.CreateNodeReply:
            type: object
            properties:
                nodeId:
                    type: string
                label:
                    type: string
                propertiesJson:
                    type: string
        api.graph.v1.CreateNodeRequest:
            type: object
            properties:
                label:
                    type: string
                propertiesJson:
                    type: string
        api.graph.v1.DeleteNodeReply:
            type: object
            properties:
                success:
                    type: boolean
        api.graph.v1.GetNodeReply:
            type: object
            properties:
                nodeId:
                    type: string
                label:
                    type: string
                propertiesJson:
                    type: string
        api.graph.v1.GetSubgraphRequest:
            type: object
            properties:
                nodeIds:
                    type: array
                    items:
                        type: string
        api.graph.v1.GraphEdge:
            type: object
            properties:
                id:
                    type: string
                source:
                    type: string
                target:
                    type: string
                type:
                    type: string
                propertiesJson:
                    type: string
        api.graph.v1.GraphNode:
            type: object
            properties:
                id:
                    type: string
                label:
                    type: string
                propertiesJson:
                    type: string
        api.graph.v1.GraphReply:
            type: object
            properties:
                nodes:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.graph.v1.GraphNode'
                edges:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.graph.v1.GraphEdge'
        api.graph.v1.UpdateNodeReply:
            type: object
            properties:
                node:
                    $ref: '#/components/schemas/api.graph.v1.GraphNode'
        api.graph.v1.UpdateNodeRequest:
            type: object
            properties:
                appId:
                    type: string
                nodeId:
                    type: string
                properties:
                    type: object
                    additionalProperties:
                        type: string
...